refactor(schema): split wire.ts into a wire/ directory by resource
wire.ts had grown to 274 lines with 8 visual section dividers marking
distinct resource domains inside one z.discriminatedUnion call. Each
domain (session/history/config/workspace/git/terminal/keep-alive) now
exports its own `as const` tuple of variant schemas from its own file
under wire/; wire/index.ts spreads them into WirePayloadSchema and keeps
WIRE_PROTOCOL_VERSION, the envelope, and parseWireMessage. agent.input/
agent.event stay inline in the index since they're a thin pass-through
of agent.ts's own contract with no wire-specific shape.

The public export surface (WIRE_PROTOCOL_VERSION, AgentHistoryList/
ReadWireOptions(Schema), WirePayload(Schema), WireMessage(Schema),
parseWireMessage) is unchanged, and `../wire` / `./wire` imports keep
resolving the same way via the directory's index.ts. Removed the now-stale
biome.json formatter override that targeted the old flat wire.ts.
